+++ packages/components-js/tests/e2e/checkbox-wrapper.e2e.ts
@@ -8,9 +8,21 @@ describe('checkbox-wrapper', () => {
     await page.setContent(`
       <p-checkbox-wrapper label="Some label">
         <input type="checkbox" name="some-name"/>
-      </p-checkbox-wrapper>`);
+      </p-checkbox-wrapper>
+    `);
     const el = await page.find('p-checkbox-wrapper >>> .p-checkbox-wrapper__fake-checkbox');
     expect(el).not.toBeNull();
+  });
+
+  it('should add aria-label to support screen readers properly', async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+      <p-checkbox-wrapper label="Some label">
+        <input type="checkbox" name="some-name"/>
+      </p-checkbox-wrapper>
+    `);
+    const input = await page.find('p-checkbox-wrapper input');
+    expect(input.getAttribute('aria-label')).toBe('Some label');
   });
 
   it('should not render label if label prop is not defined but should render if changed programmatically', async () => {