+++ packages/components-angular/projects/components-wrapper/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Divider`
 - Hover state for form elements
 
+### Fixed
+- Support label text of form elements for Screen readers
+
 ## [1.1.2] - 2020-03-17
 
 ### Changed